+++ .nix/pkgs/php-zstd/package.nix
@@ -1,0 +1,14 @@
+{ php, fetchFromGitHub }:
+
+php.buildPecl {
+  pname = "zstd";
+  version = "0.14.0";
+
+  src = fetchFromGitHub {
+    owner = "kjdev";
+    repo = "php-ext-zstd";
+    tag = "0.14.0";
+    hash = "sha256-oIbvaLYQ6Tp20Y/UEN7i1dtMnxGdMNcIjv6xRCyVYdE=";
+    fetchSubmodules = true;
+  };
+}