+++ src/core/etl/src/Flow/ETL/Function/ScalarFunctionChain.php
@@ -482,6 +482,16 @@ abstract class ScalarFunctionChain implements ScalarFunction
         return new StartsWith($this, $needle);
     }
 
+    public function stringAfter(ScalarFunction|string $needle, ScalarFunction|bool $includeNeedle = false) : self
+    {
+        return new StringAfter($this, $needle, $includeNeedle);
+    }
+
+    public function stringBefore(ScalarFunction|string $needle, ScalarFunction|bool $includeNeedle = false) : self
+    {
+        return new StringBefore($this, $needle, $includeNeedle);
+    }
+
     public function stringFold() : self
     {
         return new StringFold($this);