+++ src/core/etl/src/Flow/ETL/Loader/TransformerLoader.php
@@ -7,12 +7,19 @@ namespace Flow\ETL\Loader;
 use function Flow\ETL\DSL\{df, from_rows};
 use Flow\ETL\{FlowContext, Loader, Rows, Transformation, Transformer};
 
-final readonly class TransformerLoader implements Loader, OverridingLoader
+final readonly class TransformerLoader implements Closure, Loader, OverridingLoader
 {
     public function __construct(
         private Transformer|Transformation $transformer,
         private Loader $loader,
     ) {
+    }
+
+    public function closure(FlowContext $context) : void
+    {
+        if ($this->loader instanceof Closure) {
+            $this->loader->closure($context);
+        }
     }
 
     public function load(Rows $rows, FlowContext $context) : void