+++ src/lib/doctrine-dbal-bulk/src/Flow/Doctrine/Bulk/Dialect/Dialect.php
@@ -4,25 +4,25 @@ declare(strict_types=1);
 
 namespace Flow\Doctrine\Bulk\Dialect;
 
-use Flow\Doctrine\Bulk\{BulkData, TableDefinition};
+use Flow\Doctrine\Bulk\{BulkData, InsertOptions, TableDefinition, UpdateOptions};
 
 interface Dialect
 {
     /**
      * @param TableDefinition $table
      * @param BulkData $bulkData
-     * @param array<mixed> $insertOptions
+     * @param null|InsertOptions $options
      *
      * @return string
      */
-    public function prepareInsert(TableDefinition $table, BulkData $bulkData, array $insertOptions = []) : string;
+    public function prepareInsert(TableDefinition $table, BulkData $bulkData, ?InsertOptions $options = null) : string;
 
     /**
      * @param TableDefinition $table
      * @param BulkData $bulkData
-     * @param array<mixed> $updateOptions
+     * @param null|UpdateOptions $options
      *
      * @return string
      */
-    public function prepareUpdate(TableDefinition $table, BulkData $bulkData, array $updateOptions = []) : string;
+    public function prepareUpdate(TableDefinition $table, BulkData $bulkData, ?UpdateOptions $options = null) : string;
 }