+++ src/lib/doctrine-dbal-bulk/src/Flow/Doctrine/Bulk/QueryFactory.php
@@ -12,27 +12,17 @@ interface QueryFactory
      * @param AbstractPlatform $platform
      * @param TableDefinition $table
      * @param BulkData $bulkData
-     * @param array{
-     *  skip_conflicts?: boolean,
-     *  constraint?: string,
-     *  conflict_columns?: array<string>,
-     *  update_columns?: array<string>
-     * } $insertOptions $insertOptions
      *
      * @return string
      */
-    public function insert(AbstractPlatform $platform, TableDefinition $table, BulkData $bulkData, array $insertOptions = []) : string;
+    public function insert(AbstractPlatform $platform, TableDefinition $table, BulkData $bulkData, ?InsertOptions $options = null) : string;
 
     /**
      * @param AbstractPlatform $platform
      * @param TableDefinition $table
      * @param BulkData $bulkData
-     * @param array{
-     *  primary_key_columns?: array<string>,
-     *  update_columns?: array<string>
-     * } $updateOptions $updateOptions
      *
      * @return string
      */
-    public function update(AbstractPlatform $platform, TableDefinition $table, BulkData $bulkData, array $updateOptions = []) : string;
+    public function update(AbstractPlatform $platform, TableDefinition $table, BulkData $bulkData, ?UpdateOptions $options = null) : string;
 }